Implemented persistent high score
- Added main menu and UI elements.
- Added Back to Menu button to the game scene
- Used PlayerPrefs to save/load the best player highscore in comparison to itself and the current player.

GIT wasn't running most of the time (I have to work on that) so I don't have a lot of changes to show. I did try to use the json approach but in the end I used the KISS approach and PlayerPrefs are enough for simple single best score games.